refactor: move event utility functions to utils package
